Pinned goconvey to allow GPM usage
There are a few problems with Goconvey and GPM, specifically the fact
that GPM lets you use a project outside of the `$GOPATH`. Goconvey has
issues with this. In the commit following `cbae630` *(of the goconvey
repo)*, some symlink checks are added which cause issues. So, by pinning
to `cbae630` we can avoid some headache for now.